# docs/swagger.yaml
...
info:
  contact: {}
  description: Service for notifying users of changes to your platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Peacock API
  version: 1.6.1-dev+7aa84e8
paths:
...
  /releases/{environment}/after/{startTime}:
    get:
      consumes:
      - application/json
      description: Get releases after a specific date
      parameters:
      - description: Environment
        in: path
        name: environment
        required: true
        type: string
      - description: Start Time
        in: path
        name: startTime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get releases after a specific date
      tags:
      - release
...
swagger: "2.0"